requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
yfinance==0.2.43
ta==0.11.0
pandas==2.2.3
numpy
pydantic==2.10.4
requests
httpx>=0.27.0
stripe>=11.0.0
python-jose[cryptography]>=3.3.0
